Bug fix: transformations now outside the method
For some reason, when applying the transformations within the keep_hands_only method,
there is a low possibility of the transformations being applied on each frame individually.
This might be a problem with the Computation Graph of PyTorch since it's an indication that
the transformation were applied before the frames are stacked. In the code, however, the sequence is
stack -> transform